// src/shared/enums/features.enum.ts
export enum FeaturesEnum {
    USERS = "users",
    USERS_MANAGER = "users_manager",

    ROLES = "roles",
    SECRETARIES = "secretaries",

    GRP_ACCESS = "grp_access",
    GBP_ACCESS = "gbp_access",

    PROJECTS = "projects",

    LIST_GOALS = "lista_metas",
    STEPS = "etapas",

    PERMISSIONS = "permissions",

    PROPERTIES = "properties",
    PROPERTIES_VALUES = "properties_values",
    PROPERTIES_TESTADAS = "properties_testadas",
    CONFRONTANTE = "confrontante",
    INFO_CONFRONTANTE = "info_confrontante",
    IMOVEL_CONFRONTANTE = "imovel_confrontante",

    BCI = "bci",
    IPTU = "iptu",
    CONSULTA_PREVIA = "consulta_previa",

    LAYERS = "layers",
    LAYERS_MANAGER = "layers_manager",
    LAYERS_GROUPS = "layers_groups",
    LAYERS_STYLES = "layers_styles",

    LOGS = "logs",
}